Refactor to have transaction model to handle the query of transactions
