+++ src/api/gateway.go
@@ -84,9 +84,8 @@ type Visorer interface {
 	GetAllUnconfirmedTransactionsVerbose() ([]visor.UnconfirmedTransaction, [][]visor.TransactionInput, error)
 	GetTransaction(txid cipher.SHA256) (*visor.Transaction, error)
 	GetTransactionWithInputs(txid cipher.SHA256) (*visor.Transaction, []visor.TransactionInput, error)
-	// GetTransactions(flts []visor.TxFilter) ([]visor.Transaction, error)
-	GetTransactions(flts []visor.TxFilter, page *historydb.PageIndex) ([]visor.Transaction, uint64, error)
-	GetTransactionsWithInputs(flts []visor.TxFilter, page *historydb.PageIndex) ([]visor.Transaction, [][]visor.TransactionInput, uint64, error)
+	GetTransactions(flts []visor.TxFilter, page *visor.PageIndex) ([]visor.Transaction, uint64, error)
+	GetTransactionsWithInputs(flts []visor.TxFilter, page *visor.PageIndex) ([]visor.Transaction, [][]visor.TransactionInput, uint64, error)
 	AddressesActivity(addrs []cipher.Address) ([]bool, error)
 	GetWalletUnconfirmedTransactions(wltID string) ([]visor.UnconfirmedTransaction, error)
 	GetWalletUnconfirmedTransactionsVerbose(wltID string) ([]visor.UnconfirmedTransaction, [][]visor.TransactionInput, error)
