+++ src/api/transaction_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/SkycoinProject/skycoin/src/readable"
 	"github.com/SkycoinProject/skycoin/src/testutil"
 	"github.com/SkycoinProject/skycoin/src/visor"
-	"github.com/SkycoinProject/skycoin/src/visor/historydb"
 )
 
 func createUnconfirmedTxn(t *testing.T) visor.UnconfirmedTransaction {
@@ -1281,7 +1280,7 @@ func TestGetTransactions(t *testing.T) {
 
 				return true
 			})
-			var pageIndex *historydb.PageIndex
+			var pageIndex *visor.PageIndex
 			gateway.On("GetTransactions", matchFunc, pageIndex).Return(tc.getTransactionsResponse, uint64(0), tc.getTransactionsError)
 			gateway.On("GetTransactionsWithInputs", matchFunc, pageIndex).Return(tc.getTransactionsVerboseResponse.Transactions,
 				tc.getTransactionsVerboseResponse.Inputs, uint64(0), tc.getTransactionsVerboseError)
